Fetch cloud data using public url
